requirements.txt
langgraph==0.2.50
langchain==0.3.7
langchain-anthropic==0.2.4
anthropic==0.39.0
pydantic==2.10.2
python-dotenv==1.0.0
chainlit==1.3.1
typing-extensions==4.12.2
pytest==8.3.3
pytest-asyncio==0.24.0
